funciones de prawtosv5 para get y users

+++ src/stadistics/stadistics.h
@@ -7,16 +7,16 @@
 #include <stdint.h>
 
 typedef struct Stadistics{
-    uint64_t totalConnections;
+    uint16_t totalConnections;
     uint16_t concurrentConnections;
 
-    uint64_t bytesSent;
+    uint32_t bytesSent;
 }Stadistics;
 
 void stadistics_init(void);
 uint16_t get_concurrent_connections();
-uint64_t get_total_connections();
-uint64_t get_bytes_sent();
+uint16_t get_total_connections();
+uint32_t get_bytes_sent();
 void stadistics_increase_concurrent(void);
 void stadistics_decrease_concurrent();
 void stadistics_increase_bytes_sent(uint64_t bytes);
